refactor: reestructure project
