+++ src/database/index.js
@@ -1,7 +1,7 @@
 /**
  * @file Contains all database things.
  * @author Manuel Cabral
- * @version 0.0.2
+ * @version 0.0.3
  */
 
 // required modules
@@ -11,7 +11,17 @@ const URL = require('../config').DB.URL
 // setting up
 const connectDatabase = async () => {
 	await mongoose.connect(URL)
-	console.log(`Connected to ${URL} database`)
+	console.log(`Connected to ${URL} database successfully.`)
 }
 
-module.exports = connectDatabase
+module.exports = {
+	connectDatabase,
+
+	// export all models
+	models: {
+		User: require('./models/user'),
+		Message: require('./models/message'),
+		Chat: require('./models/chat'),
+		Token: require('./models/token'),
+	},
+}
